# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "EduMind MVP 0.3 API contract",
    "version": "0.3.0",
    "description": "MVP 0.1/0.2已实现接口及MVP 0.3拟实现契约；新增路径仅为T004设计，运行端点由T012–T028交付。JSON是YAML 1.2子集。"
  },
  "servers": [{"url": "/"}],
...
  "paths": {
...
    "/api/profile/me": {
      "get": {
        "tags": ["Learning"],
        "summary": "读取当前用户的最新画像版本",
        "operationId": "getMyProfile",
        "security": [{"sessionCookie": []}],
        "responses": {
          "200": {"description": "仅当前用户的最新画像快照。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StudentProfile"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "patch": {
        "tags": ["Learning"],
        "summary": "手动修正当前画像的白名单字段",
        "description": "只接受 `ProfileCorrectionRequest` 中的可修正字段。服务端从 Cookie 推导 owner，忽略任何客户端用户标识；每次有效修正创建带 `manual_correction` 证据的新不可变画像版本。",
        "operationId": "correctMyProfile",
        "security": [{"sessionCookie": []}],
        "parameters": [{"$ref": "#/components/parameters/CsrfToken"}, {"$ref": "#/components/parameters/IdempotencyKey"}, {"$ref": "#/components/parameters/IfMatchProfileVersion"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProfileCorrectionRequest"}}}},
        "responses": {
          "200": {"description": "已创建或重放画像版本。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StudentProfile"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/CsrfFailed"},
          "409": {"$ref": "#/components/responses/ProfileVersionConflict"},
          "422": {"$ref": "#/components/responses/ValidationError"}
        }
      }
    },
    "/api/profile/me/history": {
      "get": {
        "tags": ["Learning"],
        "summary": "读取当前 owner 的画像版本历史",
        "description": "按版本降序稳定返回去敏画像快照；不返回其他 owner、完整原始行为或 Provider 输入。",
        "operationId": "getMyProfileHistory",
        "security": [{"sessionCookie": []}],
        "parameters": [{"$ref": "#/components/parameters/HistoryLimit"}, {"$ref": "#/components/parameters/ProfileBeforeVersion"}],
        "responses": {
          "200": {"description": "画像历史页。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProfileHistoryPage"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      }
    },
    "/api/profile/events": {
      "post": {
        "tags": ["Learning"],
        "summary": "记录允许影响画像的学习行为",
        "description": "仅接受白名单行为摘要；测验作答必须使用专用测验提交端点，客户端不能写入得分、掌握状态或画像字段。事件先持久化，再以最小摘要尝试画像更新；Provider 失败不回滚事件。此轻量事件不直接更新掌握度。",
        "operationId": "recordProfileEvent",
        "security": [{"sessionCookie": []}],
        "parameters": [{"$ref": "#/components/parameters/CsrfToken"}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProfileEventRequest"}}}},
        "responses": {
          "202": {"description": "事件已持久化，并返回本次画像更新状态。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProfileEventReceipt"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/CsrfFailed"},
          "409": {"$ref": "#/components/responses/IdempotencyConflict"},
          "422": {"$ref": "#/components/responses/ValidationError"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {"sessionCookie": {"type": "apiKey", "in": "cookie", "name": "edumind_session", "description": "Secure、HttpOnly、SameSite=Lax、Host-only Cookie，2h 绝对有效期；数据库只存令牌哈希。"}},
    "parameters": {
      "CsrfToken": {"name": "X-CSRF-Token", "in": "header", "required": true, "schema": {"type": "string", "minLength": 64, "maxLength": 64}, "description": "写请求必须提供；服务端同时验证同源 Origin。"},
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 16, "maxLength": 128}, "description": "客户端为一次创建操作生成的随机键；与认证身份和请求摘要绑定。"},
...
      "HistoryLimit": {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 50, "default": 20}},
      "ProfileBeforeVersion": {"name": "before_version", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1}},
      "IfMatchProfileVersion": {"name": "If-Match-Profile-Version", "in": "header", "required": true, "schema": {"type": "integer", "minimum": 1}, "description": "客户端读取到的最新画像版本；不匹配时返回 PROFILE_VERSION_CONFLICT。"},
...
    },
    "responses": {
...
      "Unauthorized": {"description": "会话缺失、失效或不属于当前用户。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "CsrfFailed": {"description": "Origin 或 CSRF token 校验失败。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "NotFound": {"description": "资源不存在，或不属于当前用户。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "ValidationError": {"description": "请求未通过 schema 校验。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "IdempotencyConflict": {"description": "同一幂等键被用于不同请求摘要。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "ProfileVersionConflict": {"description": "画像版本已被其他有效更新推进；客户端必须重新读取后再提交。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
    },
    "schemas": {
...
      "Identifier": {"type": "string", "pattern": "^[A-Za-z0-9_-]{8,64}$"},
...
      "StudentProfile": {"type": "object", "required": ["id", "version", "updated_at"], "properties": {"id": {"$ref": "#/components/schemas/Identifier"}, "version": {"type": "integer", "minimum": 1}, "initial_query": {"type": ["string", "null"]}, "professional_background": {"type": ["object", "null"]}, "knowledge_base": {"type": ["object", "null"]}, "cognitive_style": {"type": ["object", "null"]}, "learning_goals": {"type": ["object", "null"]}, "error_preferences": {"type": ["array", "null"]}, "engineering_preference": {"type": ["object", "null"]}, "extended_dimensions": {"type": ["object", "null"]}, "evidence": {"type": ["object", "null"]}, "updated_at": {"type": "string", "format": "date-time"}}},
...
      "ProfileCorrectionRequest": {"type": "object", "additionalProperties": false, "minProperties": 1, "properties": {"professional_background": {"type": "object"}, "learning_goals": {"type": "object"}, "error_preferences": {"type": "array"}, "engineering_preference": {"type": "object"}, "cognitive_style": {"type": "object", "additionalProperties": false, "required": ["preference_persona"], "properties": {"preference_persona": {"type": "string", "enum": ["performance", "engineering", "academic"]}}}}},
      "ProfileHistoryPage": {"type": "object", "additionalProperties": false, "required": ["items", "next_before_version"], "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/StudentProfile"}}, "next_before_version": {"type": ["integer", "null"], "minimum": 1}}},
      "ProfileEventRequest": {"type":"object","additionalProperties":false,"required":["event_type","knowledge_node_id"],"properties":{"event_type":{"type":"string","enum":["hint_used","reexplanation_requested","resource_selected","explicit_feedback"]},"knowledge_node_id":{"$ref":"#/components/schemas/KnowledgeNodeId"},"learning_unit_id":{"$ref":"#/components/schemas/Identifier"},"scene_id":{"$ref":"#/components/schemas/Identifier"},"action":{"type":"string","enum":["hint_level_1","hint_level_2","simpler","deeper","different_example","code","exercise","too_easy","too_hard","liked_explanation","skip","mark_known","explanation","animation"]}}},
      "ProfileEventReceipt": {"type": "object", "additionalProperties": false, "required": ["evidence_id", "event_type", "profile_update_status"], "properties": {"evidence_id": {"$ref": "#/components/schemas/Identifier"}, "event_type": {"type": "string"}, "profile_update_status": {"type": "string", "enum": ["updated", "no_change", "provider_failed", "conflict"]}}},
      "KnowledgeNodeId": {"type": "string", "pattern": "^[a-z][a-z0-9-]{1,63}$"},
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "UNAUTHORIZED",
              "CSRF_FAILED",
              "NOT_FOUND",
              "VALIDATION_ERROR",
              "IDEMPOTENCY_CONFLICT",
              "PROFILE_VERSION_CONFLICT",
              "PATH_VERSION_CONFLICT",
              "GRAPH_NODE_NOT_FOUND",
              "QUIZ_SUBMISSION_INVALID",
              "PATH_UNREACHABLE",
              "ACCOUNT_CONFLICT",
              "RATE_LIMITED",
              "PROVIDER_UNAVAILABLE",
              "REVIEW_REJECTED",
              "OPERATION_CANCELED",
              "JOB_STATE_CONFLICT",
              "RENDER_UNAVAILABLE",
              "RENDER_FAILED",
              "RENDER_INTERRUPTED",
              "MEDIA_UNAVAILABLE",
              "MEDIA_NOT_READY",
              "CLASSROOM_VERSION_CONFLICT",
              "SCENE_VERSION_CONFLICT",
              "EVENT_CURSOR_INVALID",
              "EVENT_CURSOR_EXPIRED",
              "MESSAGE_CURSOR_INVALID",
              "MESSAGE_CURSOR_EXPIRED",
              "REVIEW_UNAVAILABLE",
              "CONTENT_RETRACTED",
              "ROLE_DISABLED",
              "DEBATE_UNAVAILABLE",
              "DEBATE_ALREADY_ACTIVE",
              "DEBATE_NOT_ACTIVE",
              "DEBATE_CONTEXT_UNAVAILABLE",
              "DEBATE_PUBLICATION_FAILED",
              "DEBATE_INVALID_OUTPUT",
              "DEBATE_INTERRUPTED"
            ]
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "operation_id": {
            "$ref": "#/components/schemas/Identifier"
          }
        }
      },
...
    }
  },
...
}